binder: prevent long delays in zombie reaping
The check for zombies was being done before binder_put_thread
which meant that any object that became a zombie could not
be handled right away even if safe. Also, need to trigger
zombie handling when deferring put_files_struct().

This manifested as long delays in adb completing a command
because the files_struct stayed alive longer than it should
have.

Bug: 36358874
Change-Id: Ib8f8b6cd3ea583342b9f39f1bab71a7948ec4bac
Test: tested and confirmed by bug submitter
Signed-off-by: Todd Kjos <tkjos@google.com>